refac: env: to achieve 100% test coverage
To do so two properties were added FatalHandler and OS.  FatalHandler
centralizes the error handlings which should terminate the program.  It
defaults to a panic and makes it mockable for tests.  OS is an
implementation of the OSer providing the os.* functionality needed by
Env.  This makes the used os.* functions mockable and easy to make them
fail.