fix: race conditions
The reoccurring problem here seems to be to have a self-initializing
(i.e. zero-value usable) concurrency save type which can be copied.  It
looks like that we either can have a copyable instance and a constructor
or a not copyable instance where the zero-value is usable but NOT a
copyable instance where the zero-value is usable without a constructor.