test: temp, in-memory logger fixture
To do concurrency save parallel testing every test must operate in its
own test-space.  This is implicitly achieved by the following aspects:

- If an environment's home-directory is not the system-user's home
  directory all paths for directories like the config-, caching- or
  logging directory are defined inside the environment's home directory.

- An environment's Lib-property allows to overwrite used std-lib
  functions like Chdir.  If we replace the later with a noop dummy it is
  possible to have an environment working directory which is different
  from the system's working directory.

- Is a logger's environment's home-directory a temporary directory logs
  are by default created in memory.

I.e. let t be a test's *T instance, then:

    lgg := &Logger{Env: (&env.Env{}).SetHome(t.FS().Tmp().Path())}
    lgg.Env.Lib.Chdir = func(path string) error { return nil }
    lgg.Env.ChWD(lgg.Env.Home())

sets up a *Logger-fixture which logs to memory and has an environment
Env whose Home() and WD() methods provide the test specific temporary
directory which is cleaned up at the end of the test while the system
working directory is unchanged.  NOTE leaving Chdir at the set noop
function means that it is not possible to change the working directory
using the logger's Env instance which is usually a desireable thing.