test: controller initialization of GINI command
The purpose of these three initial tests together with the controller
Init type is to guarantee a 100% test coverage of controller code.

Along with this test GINI's basic architecture was defined in form of a
classical MVC-architecture (model/view/controller).  NOTE the
MVC-pattern description in Wikipedia at the time of this writing is
plain wrong.  The whole purpose of the MVC pattern is that Model and
View can be developed in isolation of any other project code only
needing to care about their respective API.  I.e. letting the view
access the model is an anti-pattern without any benefit for the software
architecture.  Hence in GINI neither the view nor the model access any
other part of the GINI command implementation.  Instead the controller
leverages their APIs to receive, to process and to respond to user
input.